requirements.txt
appdirs==1.4.3
dj-database-url==0.4.1
Django==1.10.6
gunicorn==19.7.1
packaging==16.8
psycopg2==2.6.1
pyparsing==2.2.0
six==1.10.0
whitenoise==2.0.6
django-paypal==0.3.6
docker-compose==1.11.2
